Resolve all three merge blockers from independent review
Blocker 1 (PENDING placeholder stranding): traced store-refund.ts → refund-ledger.ts → attendees-edit.ts. A PENDING Square refund returns false from refundPayment, which made storeRefundedBooking terminal with only the payment leg (no refund_cash). The operator could not refresh because loadRefreshContext filtered quantity > 0 (excluding placeholders), and computeAttendeeRefund skipped them as unbalanced. Fixed: loadRefreshContext now prefers quantity > 0 but falls back to quantity-0 placeholders (ORDER BY (quantity > 0) DESC); computeAttendeeRefund now allows reversal for placeholder accounts (no sale legs) regardless of balance, so a settled PENDING refund can be reconciled via the refresh-payment path. Also dropped the ListingWithCount requirement (uses listing_id from the booking row, so a deleted listing does not 404 the refresh).

Blocker 2 (APPROVED): verified against the Square PaymentRefund object documentation — its status field is PENDING/COMPLETED/REJECTED/FAILED (NOT APPROVED). APPROVED is only in the separate RefundStatus enum page, not the PaymentRefund object that /v2/refunds returns. Removed APPROVED; CONFIRMED_REFUND_STATUSES is now [COMPLETED] only.

Blocker 3 (Valibot boundary): replaced the type cast with a Valibot schema (SquareRefundSchema with v.string() + v.minLength(1) for id and status, matching Square docs: id is Required with Min Length 1). v.parse runs OUTSIDE withClient so a malformed response (missing refund, empty id/status, non-string fields) throws loudly instead of normalizing to false. The client returns raw unknown; the boundary validation is in squareApi.refundPayment.

Also split retrieve-refund.test.ts (481→240) and rest-transport.test.ts (429→372) by extracting refund-payment.test.ts (247 lines) and refund-transport.test.ts (123 lines) with a shared mock-fetch helper. Added minLength boundary tests and a REJECTED status test. Exhaustive mutation on square.ts is 100% (284 killed, 2 equivalents suppressed).
